Auto-zoom on pattern selection change
