+++ lib/components/modification/editor.js
@@ -28,8 +28,8 @@ const AllModificationsMapDisplay = dynamic(
   () => import('lib/components/modifications-map/display-all'),
   {ssr: false}
 )
-const ModificationMapDisplay = dynamic(
-  () => import('lib/components/modifications-map/display'),
+const ActiveModificationMapDisplay = dynamic(
+  () => import('lib/components/modifications-map/active'),
   {ssr: false}
 )
 const ModificationMapEdit = dynamic(
@@ -150,7 +150,7 @@ export default class ModificationEditor extends React.PureComponent {
 
         {p.modification && (
           <>
-            <ModificationMapDisplay
+            <ActiveModificationMapDisplay
               feed={p.feed}
               key={`mmd-${p.modification._id}`}
               modification={p.modification}
