+++ lib/components/modifications-map/display.js
@@ -1,35 +1,24 @@
+import dynamic from 'next/dynamic'
 import React from 'react'
-import {useLeaflet} from 'react-leaflet'
-import {useSelector} from 'react-redux'
 
 import * as C from 'lib/constants'
 import colors from 'lib/constants/colors'
-import selectModificationBounds from 'lib/selectors/modification-bounds'
 
+const AddTripPatternLayer = dynamic(() => import('./add-trip-pattern-layer'))
+const AdjustSpeedLayer = dynamic(() => import('./adjust-speed-layer'))
+const PatternLayer = dynamic(() => import('./pattern-layer'))
+const RerouteLayer = dynamic(() => import('./reroute-layer'))
+const StopLayer = dynamic(() => import('./stop-layer'))
+
+/*
 import AddTripPatternLayer from './add-trip-pattern-layer'
 import AdjustSpeedLayer from './adjust-speed-layer'
 import PatternLayer from './pattern-layer'
 import RerouteLayer from './reroute-layer'
 import StopLayer from './stop-layer'
+*/
 
 export default function Display(p) {
-  const {triggerFitBounds} = p
-  const leaflet = useLeaflet()
-  const bounds = useSelector(selectModificationBounds)
-  const [fitBoundsTriggered, setFitBoundsTriggered] = React.useState(
-    triggerFitBounds
-  )
-
-  // Zoom to bounds on a trigger
-  React.useEffect(() => {
-    if (fitBoundsTriggered !== triggerFitBounds) {
-      setFitBoundsTriggered(triggerFitBounds)
-      if (bounds) {
-        leaflet.map.fitBounds(bounds)
-      }
-    }
-  }, [bounds, leaflet, triggerFitBounds, fitBoundsTriggered])
-
   const m = p.modification
   switch (m.type) {
     case C.REMOVE_TRIPS:
