+++ lib/selectors/modification-bounds.js
@@ -17,7 +17,7 @@ export default createSelector(
     let geometries = []
     if (get(m, 'type') !== ADD_TRIP_PATTERN) {
       geometries = patterns.map(p => p.geometry)
-    } else if (get(m, 'segments.length') > 1) {
+    } else if (get(m, 'segments[0].geometry.type') === 'LineString') {
       geometries = m.segments.map(s => s.geometry)
     }
     const coords = geometries.reduce(
